Refactor scripts into separate modules
