+++ index.html
@@ -5,9 +5,11 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
   <title>Leandru Martin's Portfolio</title>
   <link rel="stylesheet" href="style.css"/>
+  <link rel="icon" type="image/x-icon" href="/images/favicon_pen.png">
+
   <script defer type="module" src="script.js"></script>
   <script defer type="module" src="achievements.js"></script>
-  <link rel="icon" type="image/x-icon" href="/images/favicon_pen.png">
+  <script defer type="module" src="portfolio.js"></script>
 
   <meta name="author" content="Leandru Martin"/>
   <meta name="color-scheme" content="dark"/>
